+++ soxspipe/recipes/__init__.py
@@ -2,5 +2,5 @@
 *The pipeline recipes*
 """
 from ._base_recipe_ import _base_recipe_
-from .mbias import mbias
-from .mdark import mdark
+from .soxs_mbias import soxs_mbias
+from .soxs_mdark import soxs_mdark