+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## vx.x.x - xxdatexx
 
+* inject a 'SOXSPIPE PRE' keyword with timestamp value into prepared frames
+* check frames for 'SOXSPIPE PRE' keyword before preparing - raises exception if found
+
 ## v0.2.0 - February 27, 2020
 
 * added keyword lookups - abstracting exact keyword names from code